requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
python-dotenv
pydantic
uvloop
